+++ src/main/java/main/de/unikassel/vs/asp_modelling/ASPGenerator.java
@@ -1,5 +1,7 @@
-package main;
+package main.de.unikassel.vs.asp_modelling;
 
+import main.de.unikassel.vs.asp_modelling.syntax.Fact;
+import main.de.unikassel.vs.asp_modelling.syntax.Rule;
 import org.apache.velocity.Template;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.VelocityEngine;
