+++ src/main/java/main/de/unikassel/vs/asp_modelling/Main.java
@@ -1,4 +1,4 @@
-package main;
+package main.de.unikassel.vs.asp_modelling;
 
 public class Main {
