+++ src/main/java/main/de/unikassel/vs/asp_modelling/syntax/Body.java
@@ -1,26 +1,19 @@
-package main;
+package main.de.unikassel.vs.asp_modelling.syntax;
 
 import java.util.ArrayList;
 
-public class Head {
+public class Body {
 
     private Rule myRule;
-    private ArrayList<Predicate> predicates = new ArrayList<>();
+
+    private ArrayList<Predicate> predicates= new ArrayList<>();
 
     public Predicate withPredicate(String predicateName) {
         Predicate predicate = new Predicate();
-        predicate.setHead(this);
         predicate.setName(predicateName);
+        predicate.setBody(this);
         predicates.add(predicate);
         return predicate;
-    }
-
-    public void setRule(Rule rule){
-        this.myRule = rule;
-    }
-
-    public Rule getMyRule(){
-        return this.myRule;
     }
 
     public ArrayList<Predicate> getPredicates(){
